client/requirements.txt
picamera==1.13
Pillow==8.1.0